Restart with Django 1.11.7
I have desided to restart with the latest version of Django. The eariler project was not turning out how I was wanting it to.  I figured I would learn how to work the latest Django version as well as follow some better practices with this attempt of the Mineral Catalog. Things such as writting tests in conjunction with adding content.