Mineral Catalog V1.1
The project_init.py file has been altered to compliy better with PEP8 standards. Some of the logic has been rewritten to a new function that the file uses. A link to the mineral detail page has been added incase the user wants to go to a link for a specific mineral from the random mineral page.